+++ apps/e2e/cypress/e2e/back-office-applications/documentDownload.spec.js
@@ -8,6 +8,7 @@ import { SearchResultsPage } from '../../page_objects/searchResultsPage';
 import { validateSummaryPage, validateSummaryPageInfo } from '../../support/utils/utils';
 import { FileUploadPage } from '../../page_objects/uploadFiles';
 import { projectInformation } from '../../support/utils/createProjectInformation';
+import { isCI } from '../../support/utils/isCI';
 
 const createCasePage = new CreateCasePage();
 const applicationsHomePage = new ApplicationsHomePage();
@@ -27,24 +28,26 @@ describe('Document Download', () => {
 		cy.deleteDownloads();
 	});
 
-	it('Case Team Admin user should be able to download a document', () => {
-		cy.login(users.caseAdmin);
-		cy.visit('/');
-		const caseRef = Cypress.env('currentCreatedCase');
-		applicationsHomePage.searchFor(caseRef);
-		searchResultsPage.clickTopSearchResult();
-		validateSummaryPageInfo(projectInfo, 'complete');
-		searchResultsPage.clickLinkByText('Update project information');
-		searchResultsPage.clickLinkByText('Project documentation');
-		searchResultsPage.clickLinkByText('Project management');
-		fileUploadPage.verifyUploadButtonIsVisible();
-		fileUploadPage.uploadFile();
-		searchResultsPage.clickButtonByText('Save and continue');
-		fileUploadPage.verifyFolderDocuments(1);
-		fileUploadPage.verifyUploadIsComplete();
-		fileUploadPage.downloadFile(1);
-		cy.validateDownloadedFile('sample-doc');
-	});
+	if (!isCI()) {
+		it('Case Team Admin user should be able to download a document', () => {
+			cy.login(users.caseAdmin);
+			cy.visit('/');
+			const caseRef = Cypress.env('currentCreatedCase');
+			applicationsHomePage.searchFor(caseRef);
+			searchResultsPage.clickTopSearchResult();
+			validateSummaryPageInfo(projectInfo, 'complete');
+			searchResultsPage.clickLinkByText('Update project information');
+			searchResultsPage.clickLinkByText('Project documentation');
+			searchResultsPage.clickLinkByText('Project management');
+			fileUploadPage.verifyUploadButtonIsVisible();
+			fileUploadPage.uploadFile();
+			searchResultsPage.clickButtonByText('Save and continue');
+			fileUploadPage.verifyFolderDocuments(1);
+			fileUploadPage.verifyUploadIsComplete();
+			fileUploadPage.downloadFile(1);
+			cy.validateDownloadedFile('sample-doc');
+		});
+	}
 
 	it('Case Team user should be able to download a document', () => {
 		cy.login(users.caseTeam);
