+++ apps/e2e/cypress/e2e/back-office-applications/documentUpload.spec.js
@@ -23,9 +23,9 @@ describe('Document Upload', () => {
 		createCasePage.createCase(projectInfo);
 	});
 
-	it('Case Team Admin user should be able to upload a document to a case', () => {
+	it('Case Team user should be able to upload a document to a case', () => {
 		cy.clearAllCookies();
-		cy.login(users.caseAdmin);
+		cy.login(users.caseTeam);
 		cy.visit('/');
 		const caseRef = Cypress.env('currentCreatedCase');
 		applicationsHomePage.searchFor(caseRef);
