+++ apps/e2e/cypress/e2e/back-office-applications/updateProjectInformation.spec.js
@@ -12,6 +12,7 @@ import {
 } from '../../support/utils/utils';
 import { projectInformation } from '../../support/utils/createProjectInformation';
 import { CasePage } from '../../page_objects/casePage';
+import { isCI } from '../../support/utils/isCI';
 
 const applicationsHomePage = new ApplicationsHomePage();
 const casePage = new CasePage();
@@ -69,31 +70,33 @@ describe('Update Project Information', () => {
 		});
 	});
 
-	context('As Case Team Admin', () => {
-		let projectInfo = projectInformation();
-		let projectInfoNew = projectInformation();
+	if (!isCI()) {
+		context('As Case Team Admin', () => {
+			let projectInfo = projectInformation();
+			let projectInfoNew = projectInformation();
 
-		before(() => {
-			cy.login(users.caseAdmin);
-			createCasePage.createCase(projectInfo, true);
-		});
+			before(() => {
+				cy.login(users.caseAdmin);
+				createCasePage.createCase(projectInfo, true);
+			});
 
-		it('Should be able to update the case information', () => {
-			cy.login(users.caseAdmin);
-			cy.visit('/');
-			const caseRef = Cypress.env('currentCreatedCase');
-			applicationsHomePage.searchFor(caseRef);
-			searchResultsPage.clickTopSearchResult();
-			validateSummaryPageInfo(projectInfo, 'mandatory');
-			searchResultsPage.clickLinkByText('Update project information');
-			searchResultsPage.clickAccordionByText('Show all sections');
-			validateProjectInformation('Project information', projectInfo, 'mandatory');
-			updateProjectInformation(projectInfoNew);
-			validateProjectInformation('Project information', projectInfoNew, 'complete', true);
-			searchResultsPage.clickLinkByText('Preview and publish project');
-			validateProjectInformation('Preview and publish', projectInfoNew, 'complete', true);
-			searchResultsPage.clickButtonByText('Accept and publish project');
-			searchResultsPage.validateBannerMessage('Project page published');
+			it('Should be able to update the case information', () => {
+				cy.login(users.caseAdmin);
+				cy.visit('/');
+				const caseRef = Cypress.env('currentCreatedCase');
+				applicationsHomePage.searchFor(caseRef);
+				searchResultsPage.clickTopSearchResult();
+				validateSummaryPageInfo(projectInfo, 'mandatory');
+				searchResultsPage.clickLinkByText('Update project information');
+				searchResultsPage.clickAccordionByText('Show all sections');
+				validateProjectInformation('Project information', projectInfo, 'mandatory');
+				updateProjectInformation(projectInfoNew);
+				validateProjectInformation('Project information', projectInfoNew, 'complete', true);
+				searchResultsPage.clickLinkByText('Preview and publish project');
+				validateProjectInformation('Preview and publish', projectInfoNew, 'complete', true);
+				searchResultsPage.clickButtonByText('Accept and publish project');
+				searchResultsPage.validateBannerMessage('Project page published');
+			});
 		});
-	});
+	}
 });
