+++ apps/e2e/cypress/support/utils/isCI.js
@@ -1,0 +1,5 @@
+function isCI() {
+	return Cypress.env('isCI');
+}
+
+module.exports = { isCI };
